+++ CHANGELOG.md
@@ -21,6 +21,8 @@
   create, display and modify corosync configuration with such transports, use
   knet transport instead. This affects mainly `pcs cluster setup`, `pcs config
   update` and `pcs cluster link` commands.
+- Specifying score as a standalone value in `pcs constraint location add` and
+  `pcs constraint colocation add`. Use `score=value` instead.
 
 ### Fixed
 - Command `pcs stonith update-scsi-devices` no longer triggers unnecessary